PermutationSpliterator: use int (long is unnecessary here)

+++ src/main/java/one/util/streamex/PermutationSpliterator.java
@@ -92,7 +92,7 @@ import java.util.function.Consumer;
         if (remainingSize <= 1)
             return null;
         int[] newValue = value.clone();
-        long used = -1L; // clear bit = used position
+        int used = -1; // clear bit = used position
         long newRemainingSize = remainingSize / 2;
         long newPos = fence - (remainingSize -= newRemainingSize);
         long s = newPos;
@@ -102,7 +102,7 @@ import java.util.function.Consumer;
             s %= f;
             int idx = -1;
             while (rem >= 0) {
-                idx = Long.numberOfTrailingZeros(used >> (idx + 1)) + idx + 1;
+                idx = Integer.numberOfTrailingZeros(used >> (idx + 1)) + idx + 1;
                 rem--;
             }
             used &= ~(1 << idx);
